Added helm hook for rabbitmq job cluster wait
Job wait cluster was failing due to the field immutability which was
resulting in the manual delete of the job for every helm upgrade to be
successful. Reason being job being upgraded before the other manifest
that are required been updated. It can be avoided by using helm-hook
post-install and post-upgrade which will force the job manifest to be
applied only after all other manifest are applied. Hook annotation is
provided "5" so that the if other jobs are annotated, exporter job will
be last to created in case hooks are added to the other jobs in chart.
Also helm3_hook value is used for condition.

Change-Id: Ib83f1d4bef6300c2b76aa54f08927b74346184c7