Fix multidomain support selector
At the moment, the multidomain support selector is broken because
it always puts the value of a boolean inside a string which always
evaluates to true, which means setting it to false does nothing.

This patch drops the quotes around the templated configuration,
that way, it is taken for the literal boolean value.

Change-Id: I02c0a0ece680ecb55e83f3da5a992398c3ab6390