Remove helm status from deployment scripts
With the move to helm v3, helm status requires a namespace to be
specified, but doing so breaks helm v2 compatability. In order
to preserve our gating with both versions of helm while we make
the change from v2 to v3, this change removes the usage of helm
serve in openstack-helm's deployment scripts.

Once we fully move to helm v3, these scripts can be improved and
cleaned up to be more compatible with the new v3 syntax.

Change-Id: I02b6bbf780abf8c8bc7c1783c35d9411d25e18a8