[ceph-mon] Prevent mon-check from removing mons when down temporarily
A race condition exists that can cause the mon-check pod to delete
mons from the monmap that are only down temporarily. This sometimes
causes issues with the monmap when those mons come back up. This
change adds a check to see if the list of mons in the monmap is
larger than expected before removing anything. If not, the monmap
is left alone.

Change-Id: I43b186bf80741fc178c6806d24c179417d7f2406