[ceph-client] Fix ceph-rbd-pool deletion race
In cases where the pool deletion feature [0] is used, but the pool does
not exists, a pool is created and then subsequently deleted.

This was broken by the performance optimizations introduced with [1], as
the job is trying to delete a pool that does not exist (yet).

This change makes the ceph-rbd-pool job wait for manage_pools to finish
before trying to delete the pool.

0: https://review.opendev.org/c/792851
1: https://review.opendev.org/c/806443

Change-Id: Ibb77e33bed834be25ec7fd215bc448e62075f52a