Fix notifications for Cinder
At the moment, the Cinder usage audit runs every 5 minutes which
is excessive and causes load on the system.  Also, it defaults to
auditing an entire month which can take ages for large systems.

This patch makes it run sanely at the 5th minute of ever hour and
also runs the audit for the past hour only.

Change-Id: I59d1230fa4d33a2cf0364ade1a710e65ef449057
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>