Update ceph development scripts
With the move to helm v3, several of the development scripts are
either out of date or are linked to ones that are. This change
updates several of the symlinks and scripts that exist under
the ceph development directory to bring them up-to-date with
the current state of the repository.

Change-Id: I15b431d5301d47c6c83513e356c0c819d2c50963