Fix elasticsearch-data shutdown
The shutdown script for the elasticsearch-data container uses a trap
handler to run the steps outlined in the rolling restart procedure [0].
However, when trying to kill the elasticsearch process (step 3), the
script sends the TERM signal to itself.

The traps are handled recursively, causing the entire termination grace
period to be exhausted before the pod is finally removed.

This change updates the trap handler to terminate the child process(es)
instead, and wait for their completion.

0: https://www.elastic.co/guide/en/elasticsearch/reference/7.x/restart-cluster.html

Change-Id: I0c92ea5cce345cff951f044026a2179dcbd5a3e2