Remove kubeadm-aio
As part of the move to helm v3, we migrated how we setup
deploying kubernetes from a series of playbooks to using
a bash script. This had the advantage of being easier to
follow and deploy locally, as well as easier to debug
when an issue arose. While the kubeadm-aio was very useful
in the past to help lower run-times, currently it appears
that optimizations in our jobs have seem to mitigate a lot
of the issues that were lessened with the AIO.

Change-Id: If0c6a97aed4083307a9b6f5beb4ec525e8103e01