neutron: migrate IP for bridges
This patch makes a change to the Helm chart so that it migrates the
IP addresse assigned to an interface to `br-ex`.  It's assumed that
if the operator put an IP address on that interface, they likely
need it, and if they just had no IP address then it's there for L2
connectivity so nothing won't happen anyways.

Change-Id: I17dc2e532dc8b472a5c5c16ff2ec2bdcfb5bfac5