[ceph-client] Handle multiple mon versions in the pool job
The mon version check in the rbd-pool job can cause the script to
error and abort if there are multiple mon versions present in the
Ceph cluster. This change chooses the lowest-numbered major version
from the available mon versions when performing the version check
since the check is performed in order to determine the right way to
parse JSON output from a mon query.

Change-Id: I51cc6d1de0034affdc0cc616298c2d2cd3476dbb