[ceph-client] Add the ability to run Ceph commands from values
The major reason for the addition of this feature is to facilitate
an upgrade to the Pacific Ceph release, which now requires the
require-osd-release flag to be set to the proper release in order
to avoid a cluster warning scenario. Any Ceph command can be run
against the cluster using this feature, however.

Change-Id: I194264c420cfda8453c139ca2b737e56c63ef269