Start using ensure-helm role for publishing
At the moment, our publishing scripts have their own little
way of deploying Helm so it's using the old version of Helm
which is failing linting.

This updates it so that it matches how we're running it inside
our lint playbook, and it will also fix the builds since it's
also failing because of the old version of Helm.

Change-Id: I719c2dadc3ca87912234ac13d87d63e8c7b779a7