Grafana 8.5.10 with unified alerting
This PS updates Grafana to 8.5.10 and enables unified alerting
which will be the default in 9.x.

The following directories are required for unified alerting:
/var/lib/grafana/alerting
/var/lib/grafana/csv

If a user is upgrading from Grafana 7.x and wants to opt out of unified
alerting they will need to set:

[alerting]
enabled = true
[unified_alerting]
enabled = false

in grafana.ini. Additionally, to roll back, the env var

GF_DEFAULT_FORCE_MIGRATION: true

needs to be set for the grafana_run_migrator pod.

Unified alerting doc:

https://grafana.com/docs/grafana/v9.0/alerting/migrating-alerts/

Change-Id: I3a6ca005b9d9433e958802e7e978b81479a16fb8