Revert "Fix cinder host name"
This reverts commit d554d74bf0c6d61d63f92e1d8224881dd4fa726d.

This change is very problematic, because then the volumes will no longer be able to have API operations work on them, since the volumes are tied to a host, so when that host it down, things are bad.

We should keep this old (non-clean behaviour) and switch towards using Cinder clusters instead https://docs.openstack.org/cinder/latest/contributor/high_availability.html

This means configuring a few other things, but this is a really bad change and will break production deployments.  I suggest we revert ASAP.

Change-Id: Iee54ef5feca5c8bee80aded75d2fd182a431adb3