Remove placement db-migrate job
When the placement chart was originally written, there was a
need to migrate from an existing nova-placement setup.

Now that nova and placement have been separated for several
releases, there's not much use in maintaining this job. This
change removes the db-migrate functionality in the placement
chart due to issues with it in newer releases.

The chart version has been bumped to 0.3.0 to signify a
non-trivial change in functionality as well.

Change-Id: I6ff802ab8356deb5e927f414500c52b663cfa30b