[ceph-client] Correct check for too many OSDs in the pool job
The target OSD count and the final target OSD count may differ in
cases where a deployment may not include all of the hardware it is
expected to include eventually. This change corrects the check for
more OSDs running than expected to be based on the final OSD count
rather than the intermediate one to avoid false failures when the
intermediate target is exceeded and the final target is not.

Change-Id: I03a13cfe3b9053b6abc5d961426e7a8e92743808