[ceph-client] Fix OSD count checks in the ceph-rbd-pool job
This change adjusts the minimum OSD count check to be based on the
osd value, and the maxiumum OSD count check to be based on the
final_osd value. This logic supports both full deployments and
partial deployments, with the caveat that it may allow partial
deployments to over-provision storage.

Change-Id: I93aac65df850e686f92347d406cd5bb5a803659d