fix(horizon): quota container-infra in OPENSTACK_API_VERSIONS
Since most services use OPENSTACK_API_VERSIONS for the major
API version only but Magnum uses it for micro-versions, it ends
up with some decimal values.

At the moment, we have it set to 1.10 however since it's being
set as a number, it's becoming 1.1 which is too old.  The
solution is to quote it as a string instead and that is what
magnum-ui expects as well.

Change-Id: I46b5fec69d8bc353e38a6f8e6e506e2a460371f4