fix(cinder): add endpoint details for nova
This change adds the endpoint details for Nova to allow for online
resizes however a very key thing in this is that it is actually
using the Cinder credentials to talk to Nova.

OpenStack projects have historically arbitrarily decided to use
the user credentials of the _target_ service rather than the
source service which does not make sense, a mailing list discussion
seemed to have not brought up any negatives but only positives.

For the future, we can continue to do this which will simplify
our endpoints section but this is a start.

Change-Id: Ib9b500ef9a9bc34c8b64215bee57451494735573