[ceph-client] Strip any errors preceding pool properties JSON
Sometimes errors appear in the 'ceph osd pool get' output before
the JSON string. The returned string is saved and is assumed to
contain only the JSON string with the pool properties. When errors
appear in the string, pool properties are not read properly, which
can cause pools to be misconfigured. This change filters that
output so only the expected JSON string is returned. It can then be
parsed correctly.

Change-Id: I83347cc32da7e7af160b5cacc2a99de74eebebc7