Run tests for older releases periodic-weekly
Openstack releases older than Yoga are now in
extended maintenance. To reduce the CI
footprint we don't run test jobs for older
relases as part of the check/gate pipelines.

Instead we are going to run those jobs as
part of the periodic-weekly pipeline.

See the detailed description of extended maintenance
status here
https://docs.openstack.org/project-team-guide/stable-branches.html#maintenance-phases

Change-Id: Ie227e7d2dd297b6095a40f6114ef6b0a2f226790