Use multinode nodeset for compute-kit jobs
For recent releases we use 32GB nodes for compute-kit
jobs. The number of such nodes is extremely limited.
So we'd better use multinode nodesets for compute-kit
jobs.

We deploy K8s using kubeadm and then we set labels to the
K8s nodes so charts can use these labels for node selectors.
We deploy L3 agent only on the node where we run test scripts.

This is because we want test virual router to be always created
on this node. Otherwise the L2 overlay needs to be created
to emulate provider network (will be implemented later).

Glance is deployed w/o backend storage (will be fixed later).

Change-Id: Id2eb639fb67d41006940a7d7b45a865b2f1124f7