Use ubuntu-focal-32GB label for compute-kit tests
The change https://review.opendev.org/c/openstack/project-config/+/888901
introduces two new labels ubuntu-focal-32GB and ubuntu-jammy-32GB.
These labels are available in only one region (Vexxhost ca-ymq-1)
and the number of such nodes is extremely limited. So we are
going to switch compute-kit jobs to multinode nodesets ASAP and
this is in progress at the moment.

This change is a temporary solution to unblock all our
activities.

Change-Id: I2c2a54f4a90656bc6fb441b8a794744e4a636cb6