Allows keep glance-images PVC once defined
Glance image PVC contains requires images that should be careful handled
during changing glance storage when migrate out of PVC mode for glance
storage.
Which in migrate/upgrade path, should be be correctly moved and deleted
the PVC after.

On the other hand, it's also possible to accidentally changes storage
mode out of `pvc` and lose the `glance-images` PVC which is unbearable
mistake.

Once storage mode set to `pvc, we should allow that PVC to be able to
stay and ready for reuse again until it's mannually deleted.

This add flag `keep_pvc` (default to true).
Set it to true to set helm/resource-policy to keep for glance-images.
Set it to false to allow helm delete glance-images PVC when request.

Change-Id: I9d0e2a49aabf81eb2d4e00ad2a9d42125261489e