[rook-ceph] Document the procedure for migrating Ceph to Rook
This change adds documentation to describe the procedure and
example script for migrating a Ceph cluster deployed with the
legacy openstack-helm-infra Ceph charts to be deployed and managed
by Rook. Rook is the supported mechanism for running Ceph on
Kubernetes and is now the default method for deploying Ceph in
openstack-helm.

Change-Id: I43644295d67ba5efc007bbe1ee194e1dd923677b