ironic: make conductor statefulset more configurable
Made the conductor statefulset more configurable by allowing the user to
enable/disable the http or pxe containers and to modify their init
and/or runtime scripts. Allow another OS to potentially be used by
letting the user supply the HOST_OS value and FILEPATH. Simplify the code
by using the variables already defined. Include the snponly image as the
counterpart to undionly.

Change-Id: I181657bca6ace0f94f6e0ef0b6f8a5bc7a433ac3