Do not attach PVC to glance-storage-init job
When PVC is used as glance storage backend we
don't need to attach it to the storage init pod
because anyway the init script does nothing with the PVC.

On the other hand when PVC is used as the backend
there is a race condition when this PVC is attached to
the API pod and to the storage init pod. It breaks the
deployment because storage init pod never starts
and API pod waits for the storage init job to finish.

Also this PR adds the test job that deploys compute-kit
with Ceph as storageclass backend. In this job Mariadb, Rabbit,
Glance, Cinder all use Ceph while usually we test charts
with in-memory volumes. This job is heavy and we run it
periodically.

Change-Id: I7fd1a4a47472bd81cda43cb50b1e3dc448619598